refactor: move Types for articleCategory and create new createArticleCategory
