+++ src/actions/cms/articleCategories/destroy.ts
@@ -1,7 +1,7 @@
 'use server'
 import prisma from '@/prisma'
 import { createPrismaActionError } from '@/actions/error'
-import type { ExpandedArticleCategory } from './Types'
+import type { ExpandedArticleCategory } from '@/cms/articleCategories/Types'
 import type { ActionReturn } from '@/actions/Types'
 
 export async function destroyArticleCategory(id: number): Promise<ActionReturn<ExpandedArticleCategory>> {
