+++ src/app/articles/[category]/SideBar.tsx
@@ -10,7 +10,7 @@ import Link from 'next/link'
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome'
 import { faChevronUp, faX } from '@fortawesome/free-solid-svg-icons'
 import { useRouter } from 'next/navigation'
-import type { ExpandedArticleCategory } from '@/actions/cms/articleCategories/Types'
+import type { ExpandedArticleCategory } from '@/cms/articleCategories/Types'
 
 type PropTypes = {
     category: ExpandedArticleCategory
